feat: recreate logic to render form fields on MultistepForm
