+++ packages/multistep-form/src/Multistep/MultistepFormStepper.tsx
@@ -1,0 +1,66 @@
+import * as React from 'react';
+import { Button } from '@ttoss/ui';
+
+import { Form, useForm, yup, yupResolver } from '@ttoss/forms';
+import {
+  MultistepFlowMessage,
+  MultistepFlowMessageProps,
+} from './MultistepFlowMessage';
+import { MultistepQuestion } from './MultistepQuestion';
+
+export type MultistepFormStepperProps = {
+  flowMessage: MultistepFlowMessageProps;
+  onSubmit: (data: unknown) => void;
+  question: string;
+  isLastStep: boolean;
+  fields: React.ReactNode | React.ReactNode[];
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  schema?: yup.ObjectSchema<any>;
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  defaultValues?: any;
+  submitLabel: string;
+};
+
+export const MultistepFormStepper = ({
+  flowMessage,
+  fields,
+  onSubmit,
+  question,
+  submitLabel,
+  schema,
+  isLastStep,
+  defaultValues,
+}: MultistepFormStepperProps) => {
+  const formMethods = useForm({
+    resolver: schema ? yupResolver(schema) : undefined,
+    defaultValues,
+  });
+
+  return (
+    <Form
+      {...formMethods}
+      sx={{
+        display: 'flex',
+        flexDirection: 'column',
+      }}
+      onSubmit={onSubmit}
+    >
+      <MultistepFlowMessage {...flowMessage} />
+
+      <MultistepQuestion fields={fields} question={question} />
+
+      <Button
+        sx={{
+          justifyContent: 'center',
+          marginTop: '2xl',
+          marginBottom: 'xl',
+          marginX: '2xl',
+        }}
+        rightIcon={isLastStep ? undefined : 'nav-right'}
+        type="submit"
+      >
+        {submitLabel}
+      </Button>
+    </Form>
+  );
+};
