+++ packages/multistep-form/src/Multistep/MultistepHeaderLogo.tsx
